Fix some type discrepancies with the length passed into crosscorr
The types of int16_t and int were causing innaccurate values to be passed in
as the result wasn't cast correctly from one to the other

git-svn-id: svn+ssh://svn.forge.ecs.soton.ac.uk/projects/sanchp3yp@131 bc9da1e1-76ae-0410-a611-e94dc21032b9

+++ work/code/sigproc.c
@@ -25,7 +25,7 @@ void cancel(void * ring1, int16_t * ptr1in, void * ring2, int16_t * ptr2in, void
 	}
 }
 
-void crosscorr(void * ring1, int16_t * ptr1in, void * ring2, int16_t * ptr2in, double * res, int length, int maxdel)
+void crosscorr(void * ring1, int16_t * ptr1in, void * ring2, int16_t * ptr2in, double * res, int16_t length, int16_t maxdel)
 {
 	int del, i, j;
 	double mean1, mean2;
